Add media query for desktop devices
Added a media query to change the arrange of the card elements to match
the desktop design.

Added the source tag to the picture tag, with an attribute of media,
to change the image when a specific screen size is reached.

Also the clamp property was used to have a responsive font size for the
item description.