Modify item-information arrangement
The HTML content of the item information was overcomplicated,
so it was changed and simplified.

The flex property was used to let grow evenly the 2 containers in the
item card, also some spacing values for the design were modified.

To the picture element and its img within it were applied height:100%
to not have unused space .